*Implemented CORE-1274,1278,1312

+++ api/src/main/java/org/ednovo/gooru/domain/service/ScollectionServiceImpl.java
@@ -70,6 +70,7 @@ import org.ednovo.gooru.core.application.util.ResourceMetaInfo;
 import org.ednovo.gooru.core.constant.ConstantProperties;
 import org.ednovo.gooru.core.constant.ParameterProperties;
 import org.ednovo.gooru.core.exception.NotFoundException;
+import org.ednovo.gooru.core.exception.UnauthorizedException;
 import org.ednovo.gooru.domain.cassandra.service.ResourceCassandraService;
 import org.ednovo.gooru.domain.service.assessment.AssessmentService;
 import org.ednovo.gooru.domain.service.partner.CustomFieldsService;
@@ -94,6 +95,7 @@ import org.ednovo.gooru.infrastructure.persistence.hibernate.resource.ResourceRe
 import org.ednovo.gooru.infrastructure.persistence.hibernate.taxonomy.TaxonomyRespository;
 import org.ednovo.gooru.infrastructure.persistence.hibernate.taxonomy.TaxonomyStoredProcedure;
 import org.ednovo.gooru.json.serializer.util.JsonSerializer;
+import org.ednovo.gooru.security.OperationAuthorizer;
 import org.ednovo.goorucore.application.serializer.JsonDeserializer;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -194,6 +196,9 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 
 	@Autowired
 	private UserRepository userRepository;
+	
+	@Autowired
+	private OperationAuthorizer operationAuthorizer;
 
 	Logger logger = LoggerFactory.getLogger(ScollectionServiceImpl.class);
 
@@ -614,52 +619,57 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 	public void deleteCollection(String collectionId, User user) {
 		Collection collection = this.getCollectionByGooruOid(collectionId, null);
 		if (collection != null) {
-			try {
-				revisionHistoryService.createVersion(collection, SCOLLECTION_DELETE);
-			} catch (Exception ex) {
-				ex.printStackTrace();
-			}
-			try {
-				indexProcessor.index(collection.getGooruOid(), IndexProcessor.DELETE, SCOLLECTION);
-			} catch (Exception e) {
-				logger.debug(e.getMessage());
-			}
+			if(this.getOperationAuthorizer().hasUnrestrictedContentAccess(collectionId, user)){
+				try {
+					revisionHistoryService.createVersion(collection, SCOLLECTION_DELETE);
+				} catch (Exception ex) {
+					ex.printStackTrace();
+				}
+				try {
+					indexProcessor.index(collection.getGooruOid(), IndexProcessor.DELETE, SCOLLECTION);
+				} catch (Exception e) {
+					logger.debug(e.getMessage());
+				}
 
-			List<CollectionItem> collectionItems = this.getCollectionRepository().getCollectionItemByAssociation(collectionId, null, null);
-			try {
-				getEventLogs(collection.getCollectionItem(), user, collection.getCollectionItem().getCollection().getCollectionType());
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
+				List<CollectionItem> collectionItems = this.getCollectionRepository().getCollectionItemByAssociation(collectionId, null, null);
+				try {
+					getEventLogs(collection.getCollectionItem(), user, collection.getCollectionType());
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
 
-			for (CollectionItem item : collectionItems) {
-				this.deleteCollectionItem(item.getCollectionItemId(), user);
-				if (item.getAssociatedUser() != null && !item.getAssociatedUser().getPartyUid().equals(user.getPartyUid())) {
-					this.redisService.bulkKeyDelete("v2-organize-data-" + item.getAssociatedUser().getPartyUid() + "*");
+				for (CollectionItem item : collectionItems) {
+					this.deleteCollectionItem(item.getCollectionItemId(), user);
+					if (item.getAssociatedUser() != null && !item.getAssociatedUser().getPartyUid().equals(user.getPartyUid())) {
+						this.redisService.bulkKeyDelete("v2-organize-data-" + item.getAssociatedUser().getPartyUid() + "*");
+					}
 				}
-			}
-			if (collection != null && collection.getUser() != null && collection.getSharing().equalsIgnoreCase(PUBLIC)) {
-				UserSummary userSummary = this.getUserRepository().getSummaryByUid(collection.getUser().getPartyUid());
-				if (userSummary != null && userSummary.getCollections() != null) {
-					userSummary.setCollections(userSummary.getCollections() <= 0 ? 0 : (userSummary.getCollections() - 1));
-					this.getUserRepository().save(userSummary);
+				if (collection != null && collection.getUser() != null && collection.getSharing().equalsIgnoreCase(PUBLIC)) {
+					UserSummary userSummary = this.getUserRepository().getSummaryByUid(collection.getUser().getPartyUid());
+					if (userSummary != null && userSummary.getCollections() != null) {
+						userSummary.setCollections(userSummary.getCollections() <= 0 ? 0 : (userSummary.getCollections() - 1));
+						this.getUserRepository().save(userSummary);
+					}
+					this.getUserRepository().flush();
 				}
-				this.getUserRepository().flush();
-			}
-			this.getCollectionRepository().remove(Collection.class, collection.getContentId());
-			for (CollectionItem item : collectionItems) {
-				Collection parentCollection = item.getCollection();
-				if (parentCollection.getCollectionType().equals(FOLDER)) {
-					updateFolderSharing(parentCollection.getGooruOid());
+				this.getCollectionRepository().remove(Collection.class, collection.getContentId());
+				for (CollectionItem item : collectionItems) {
+					Collection parentCollection = item.getCollection();
+					if (parentCollection.getCollectionType().equals(FOLDER)) {
+						updateFolderSharing(parentCollection.getGooruOid());
+					}
 				}
+			} else {
+				throw new UnauthorizedException("user don't have permission ");
 			}
+			
 		} else {
 			throw new NotFoundException(generateErrorMessage(GL0056, _COLLECTION));
 		}
 		this.redisService.bulkKeyDelete("v2-organize-data-" + user.getPartyUid() + "*");
 
 	}
-
+	
 	@Override
 	public List<Collection> getCollections(Map<String, String> filters, User user) {
 		return this.getCollectionRepository().getCollections(filters, user);
@@ -829,29 +839,34 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 	@Override
 	public void deleteCollectionItem(String collectionItemId, User user) {
 		CollectionItem collectionItem = this.getCollectionRepository().getCollectionItemById(collectionItemId);
-		if (collectionItem != null) {
-			try {
-				getEventLogs(collectionItem, user, collectionItem.getCollection().getCollectionType());
-			} catch (JSONException e1) {
-				e1.printStackTrace();
-			}
-			Collection collection = collectionItem.getCollection();
-			this.getCollectionRepository().remove(CollectionItem.class, collectionItem.getCollectionItemId());
-
-			collectionItem.getCollection().setLastUpdatedUserUid(user.getPartyUid());
-			collectionItem.getCollection().setLastModified(new Date(System.currentTimeMillis()));
-			collectionItem.getCollection().setItemCount((collectionItem.getCollection().getItemCount() == null || (collectionItem.getCollection().getItemCount() != null && collectionItem.getCollection().getItemCount() == 0)) ? 0 : collectionItem.getCollection().getItemCount() - 1);
-			reOrderCollectionItems(collection, collectionItemId);
-			try {
-				if (collectionItem.getResource().getResourceType() != null && collectionItem.getResource().getResourceType().getName().equalsIgnoreCase(ResourceType.Type.SCOLLECTION.getType())) {
-					indexProcessor.index(collectionItem.getResource().getGooruOid(), IndexProcessor.DELETE, SCOLLECTION);
-				} else {
-					indexProcessor.index(collectionItem.getResource().getGooruOid(), IndexProcessor.INDEX, RESOURCE);
+		if (collectionItem != null && collectionItem.getResource() != null) {
+			if(this.getOperationAuthorizer().hasUnrestrictedContentAccess(collectionItem.getResource().getGooruOid(), user)){
+				try {
+					getEventLogs(collectionItem, user, collectionItem.getCollection().getCollectionType());
+				} catch (JSONException e1) {
+					e1.printStackTrace();
 				}
-				indexProcessor.index(collectionItem.getCollection().getGooruOid(), IndexProcessor.INDEX, SCOLLECTION);
-			} catch (Exception e) {
-				logger.debug(e.getMessage());
+				Collection collection = collectionItem.getCollection();
+				this.getCollectionRepository().remove(CollectionItem.class, collectionItem.getCollectionItemId());
+
+				collectionItem.getCollection().setLastUpdatedUserUid(user.getPartyUid());
+				collectionItem.getCollection().setLastModified(new Date(System.currentTimeMillis()));
+				collectionItem.getCollection().setItemCount((collectionItem.getCollection().getItemCount() == null || (collectionItem.getCollection().getItemCount() != null && collectionItem.getCollection().getItemCount() == 0)) ? 0 : collectionItem.getCollection().getItemCount() - 1);
+				reOrderCollectionItems(collection, collectionItemId);
+				try {
+					if (collectionItem.getResource().getResourceType() != null && collectionItem.getResource().getResourceType().getName().equalsIgnoreCase(ResourceType.Type.SCOLLECTION.getType())) {
+						indexProcessor.index(collectionItem.getResource().getGooruOid(), IndexProcessor.DELETE, SCOLLECTION);
+					} else {
+						indexProcessor.index(collectionItem.getResource().getGooruOid(), IndexProcessor.INDEX, RESOURCE);
+					}
+					indexProcessor.index(collectionItem.getCollection().getGooruOid(), IndexProcessor.INDEX, SCOLLECTION);
+				} catch (Exception e) {
+					logger.debug(e.getMessage());
+				}
+			} else {
+				throw new UnauthorizedException("user don't have permission ");
 			}
+			
 		} else {
 			throw new NotFoundException(generateErrorMessage(GL0056, _COLLECTION_ITEM));
 
@@ -2425,8 +2440,8 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 	public void getEventLogs(CollectionItem collectionItem, User user, String collectionType) throws JSONException {
 		SessionContextSupport.putLogParameter(EVENT_NAME, "item.delete");
 		JSONObject context = SessionContextSupport.getLog().get("context") != null ? new JSONObject(SessionContextSupport.getLog().get("context").toString()) : new JSONObject();
-		context.put("parentGooruId", collectionItem.getCollection() != null ? collectionItem.getCollection().getGooruOid() : null);
-		context.put("contentGooruId", collectionItem.getResource() != null ? collectionItem.getResource().getGooruOid() : null);
+		context.put("parentGooruId", collectionItem != null && collectionItem.getCollection() != null ? collectionItem.getCollection().getGooruOid() : null);
+		context.put("contentGooruId", collectionItem != null && collectionItem.getResource() != null ? collectionItem.getResource().getGooruOid() : null);
 		SessionContextSupport.putLogParameter("context", context.toString());
 		JSONObject payLoadObject = SessionContextSupport.getLog().get("payLoadObject") != null ? new JSONObject(SessionContextSupport.getLog().get("payLoadObject").toString()) : new JSONObject();
 		if (collectionType != null && collectionItem != null) {
@@ -2528,6 +2543,14 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 
 	public FeedbackService getFeedbackService() {
 		return feedbackService;
+	}
+
+	public OperationAuthorizer getOperationAuthorizer() {
+		return operationAuthorizer;
+	}
+
+	public void setOperationAuthorizer(OperationAuthorizer operationAuthorizer) {
+		this.operationAuthorizer = operationAuthorizer;
 	}
 
 	public AsyncExecutor getAsyncExecutor() {
